requirements.txt
numpy==2.2.4
pillow==12.1.0
pycryptodome==3.23.0
